feat(extension-api): expose shared Logger contract and configure it in ThorApplication
